VocalTrack: use compile time index for load and scan
When scan or loading a track from .mid, the actual track index is known ahead of time in terms of which trackname is applies to. Therefore, we can use templated indexing instead of a function parameter to generate the appropriate code. No need to constantly check the index value at run time.