SongCache: replace list of Songs w/ vector of Song pointers
The original intention of using std::list was so iterator would not be invalidated when adding new Song objects. Besides the obvious downside of no random access, this also made it so the song (valid or not) HAD to be added to the list.

By moving to pointers, each thread can check validity before adding the object to the final vector. This simplifies the algorithm for removing duplicate song hashes.