Unicode usage rework
Most instances of strings based in unicode do not require the casing functionality of the UnicodeString class. Therefore, holding those as basic u32strings instead of instances of the class is far more efficient.

Doing so just requires utilizing the static encode and decoding functions of the UnicodeString class.

Note: the exception to this is anything that could possibly be used as a menu UI element (ex: song select screen names/categories or section names in a practice mode)