Song: rework how noteTracks are created globally
Instantiate each track separately and in contiguous memory in a static collected structure. In that structure, have a separate array that holds pointers to each separate track in order. This style will still have the same ability to access as an array while being much easier to understand on a track-by-track basis. This most notably effects .cht & .mid as reinterpret_cast is no longer necessary.