Ini: class rework
No longer hold every single possible modifier as member variables of the class. Instead, each instance will hold a vector of all parsed modifiers from a file. This leads to a significant decrease in total memory usage by the type.

To achieve this, the modifier map has been reconstructed to hold pairs of string_views and a functions that create new Modifier objects. While this only applies to IniFile, in the future, this will be expanded to all iterations of TextTraversal in general.

With this change, Modifiers no longer need conditional statements for writing their values nor need "default" values.  Modifiers can be just be removed or added depending on the situation.

Note: instances of IniFile will always have the main songInfo modifiers in its vector.