TxtFileModifier: funny custom "variant" implementation go brr
Me no likey vtable stuff.

For real though, one of the main reasons why I wanted to swap to std::variant was to do away with some of the virtual stuff. However, with how std::variant is generalized, virtualization was still a part of the system. Since the possible modifier types used are known, however, it was very possible to just customize txtFileModifier to provide the same functionality, but in a more defined non-virtual fashion.

This does indeed provide a performance bump, particularly in scanning.